refac: 서비스 클래스 책임 분리
