+++ src/main/java/com/innercircle/project_one/survey/api/service/SurveyAnswerService.java
@@ -1,0 +1,110 @@
+package com.innercircle.project_one.survey.api.service;
+
+import com.innercircle.project_one.survey.api.dto.SurveySubmitDTO;
+import com.innercircle.project_one.survey.api.repository.*;
+import com.innercircle.project_one.survey.common.ApiResponse;
+import com.innercircle.project_one.survey.common.SuccessResponse;
+import com.innercircle.project_one.survey.common.SurveyObjectDataType;
+import com.innercircle.project_one.survey.domain.*;
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Service
+@AllArgsConstructor
+public class SurveyAnswerService {
+
+    private final SurveyService surveyService;
+    private final SurveyObjectAnswerRepository surveyObjectAnswerRepository;
+
+
+    @Transactional
+    public ApiResponse submitSurveyResponse(Long surveyId, SurveySubmitDTO surveySubmitDTO) {
+
+        Survey survey = surveyService.findSurvey(surveyId);
+        isSubmitable(survey, surveySubmitDTO);
+
+        List<SurveyObject> surveyObjects = survey.getSurveyObjects();
+        survey.sortSurveyObjects();
+
+        List<SurveySubmitDTO.SurveySubmitObject> objects = surveySubmitDTO.objects();
+
+        for (int i = 0; i < objects.size(); i++) {
+            SurveySubmitDTO.SurveySubmitObject submitObject = objects.get(i);
+
+            List<SurveyObjectAnswer> answers = createAnswer(surveyObjects.get(i), submitObject);
+            for (SurveyObjectAnswer answer : answers) {
+                surveyObjectAnswerRepository.save(answer);
+            }
+        }
+
+        return new SuccessResponse<>("설문조사 응답이 제출되었습니다.");
+    }
+
+
+    private List<SurveyObjectAnswer> createAnswer(SurveyObject surveyObject, SurveySubmitDTO.SurveySubmitObject requestObject) {
+        List<SurveyObjectAnswer> answers = new ArrayList<>();
+        switch (SurveyObjectDataType.of(requestObject.type())) {
+            case TEXT, RICH_TEXT -> {
+                String content = ((SurveySubmitDTO.SurveySubmitObject.StringContent) requestObject.content()).getContent();
+                answers.add(StringSurveyObjectAnswer.builder()
+                        .surveyObject(surveyObject)
+                        .answer(content)
+                        .build());
+            }
+            case RADIO -> {
+                String selectedElement = ((SurveySubmitDTO.SurveySubmitObject.ElementContent) requestObject.content()).getSelectedElement();
+                answers.add(ElementSurveyObjectAnswer.builder()
+                        .surveyObject(surveyObject)
+                        .elementObject(new ElementObject(1, selectedElement, surveyObject))
+                        .build());
+            }
+            case CHECK_BOX -> {
+                List<String> selectedElements = ((SurveySubmitDTO.SurveySubmitObject.CheckBoxContent) requestObject.content()).getSelectedElements();
+                for (int i = 0; i < selectedElements.size(); i++) {
+                    answers.add(ElementSurveyObjectAnswer.builder()
+                            .surveyObject(surveyObject)
+                            .elementObject(new ElementObject(i + 1, selectedElements.get(i), surveyObject))
+                            .build());
+                }
+            }
+            default -> throw new IllegalArgumentException("적절한 입력 타입이 아닙니다.");
+        }
+        return answers;
+    }
+
+    private void isSubmitable(Survey survey, SurveySubmitDTO surveySubmitDTO){
+
+        if(survey.getSurveyVersion().getVersion() != surveySubmitDTO.version()) {
+            throw new IllegalArgumentException("버전이 일치하지 않습니다.");
+        }
+
+        survey.sortSurveyObjects();
+        List<SurveyObject> savedSurveyObjects = survey.getSurveyObjects();
+        List<SurveySubmitDTO.SurveySubmitObject> submitSurveyObjects = surveySubmitDTO.objects();
+
+        if (savedSurveyObjects.size() != submitSurveyObjects.size()) {
+            throw new IllegalArgumentException("목록이 일치하지 않습니다.");
+        }
+
+        for (int i = 0; i < savedSurveyObjects.size(); i++){
+            SurveyObject surveyObject = savedSurveyObjects.get(i);
+            SurveySubmitDTO.SurveySubmitObject submitSurveyObject = submitSurveyObjects.get(i);
+
+//            if(!Objects.equals(surveyObject.getId(), submitSurveyObject.id())){
+//                throw new IllegalArgumentException("목록 순서가 일치하지 않습니다.");
+//            }
+
+            if (surveyObject.getType() != SurveyObjectDataType.of(submitSurveyObject.type())) {
+                throw new IllegalArgumentException("목록 타입이 일치하지 않습니다.");
+            }
+
+        }
+
+    }
+
+
+}
